requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
wikipedia==1.4.0
numexpr==2.8.7
chromadb==0.4.24 